pyproject.toml
[tool.poetry]
name = "assignment-2"
version = "0.1.0"
description = ""
authors = ["spicadust <spicastre@gmail.com>"]
readme = "README.md"
packages = [
    { include = "guided_diffusion" },
]
[tool.poetry.dependencies]
python = "^3.12"
blobfile = ">=1.0.5"
fastai = "^2.7.18"
notebook = "^7.2.2"
python-dotenv = "*"
torch = "*"
tqdm = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"